create useProductDetailData hook
